# web/doc/openapi.yaml
openapi: 3.0.3
info:
  title: R-Clash API Specification
  description: Foundation API for the modern R-Clash client. Adheres to Action-oriented and Smart Merge principles.
  version: 1.5.0
servers:
  - url: /api
    description: Production API Server
security:
  - bearerAuth: []
paths:
...
  /proxies/{group}:
    put:
      summary: 切换代理组选中的出口
      parameters:
        - name: group
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, description: "选中的节点名称" }
      responses:
        '200':
          description: 切换成功
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: 使用 RWEB_CLASH_API_TOKEN；非回环监听时必须配置，本地回环且未配置令牌时可省略。
...